Refactored utility function out of platform gulp files.
